Code refactoring - All classes are now in corresponding file - main file added - *.svg is default format in saving filedialog
