+++ Main.py
@@ -1,0 +1,6 @@
+from StandMaker import StandMaker
+
+
+if __name__ == '__main__':
+    app = StandMaker()
+    app.mainloop()
